src/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asynco==0.0.5
asyncpg==0.30.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
prometheus_client==0.21.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
redis==5.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
